new state of redux configuration
