+++ src/store/actions/background.js
@@ -1,0 +1,6 @@
+export function addBackground(background) {
+  return {
+    type: '@background/ADD_BACKGROUND',
+    payload: { background },
+  };
+}
