+++ src/store/actions/brazilStates.js
@@ -1,0 +1,6 @@
+export function updateBrazilStates(brazilStates) {
+  return {
+    type: '@brazilStates/UPDATE_BRAZIL_STATES',
+    payload: { brazilStates },
+  };
+}
