+++ src/store/actions/index.js
@@ -1,0 +1,5 @@
+export * from './background';
+export * from './brazilStates';
+export * from './location';
+export * from './unit';
+export * from './weather';
