+++ src/store/actions/location.js
@@ -1,0 +1,6 @@
+export function updateLocation(location) {
+  return {
+    type: '@location/UPDATE_LOCATION',
+    payload: { location },
+  };
+}
