+++ src/store/actions/unit.js
@@ -1,0 +1,6 @@
+export function updateUnit(unit) {
+  return {
+    type: '@unit/UPDATE_UNIT',
+    payload: { unit },
+  };
+}
