+++ src/store/actions/weather.js
@@ -1,0 +1,6 @@
+export function updateWeather(weather) {
+  return {
+    type: '@weather/UPDATE_WEATHER',
+    payload: { weather },
+  };
+}
