+++ src/store/index.js
@@ -1,6 +1,6 @@
 import { createStore } from 'redux';
 
-import rootReducer from './module/rootReducer';
+import rootReducer from './reducers';
 
 const store = createStore(rootReducer);
