--- src/store/module/rootReducer.js
@@ -1,5 +1,0 @@
-import { combineReducers } from 'redux';
-
-import main from './main/reducer';
-
-export default combineReducers({ main });
